requirements.txt
flask
flask-socketio
cryptography
spotipy
python-dotenv
lrcup
dateparser
asyncio
